+++ backend/src/main/java/com/example/journalsystem/Repository/UserRepository.java
@@ -12,7 +12,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
 
     List<User> getUserByEmail(String email);
 
-    Optional<User> getUserById(Long id);
+    Optional<User> getUserById(int id);
 
     List<User> findAll();
