+++ backend/src/main/java/com/example/journalsystem/models/Observation/Observation.java
@@ -16,11 +16,11 @@ public class Observation {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private int id;
     private String name;
     private String description;
     private LocalDate observationDate;
-    private Long userId;
+    private int userId;
 
     @ManyToOne
     @JoinColumn(name = "user_id", nullable = false) // Definierar foreign key
